// src/generated/flight-reshop.ts
export interface components {
    schemas: {
// ...
        FlightReshopRequest: {
            /**
             * @description The pseudo city code of the target destination for which the flight reshop is requested.
             * @example AAA
             */
            targetPcc?: string;
            /**
             * @description The booking reference ID as shown in the source supplier/vendor system. In the traditional ATPCO distribution model, this corresponds to the PNR Locator value.
             * @example GLEBNY
             */
            bookingId?: string;
            /** @description Lists information about all individual journeys to search for. */
            journeys: components["schemas"]["RequestedJourney"][];
            /** @description Lists information about the travelers. */
            travelers?: components["schemas"]["RequestedTraveler"][];
            /** @description Contains qualifiers that allow you to modify the requested trip. */
            route?: components["schemas"]["RequestedRoute"];
            /** @description Contains qualifiers that allow you to define date flexibility for the requested trip. By default, multiple offers may be returned for the main travel date, with a maximum of one offer per date for the remaining dates within the flexibility range. */
            departureDateFlexibility?: components["schemas"]["DateFlexibility"];
            /** @description Contains carrier filters for airlines that the API is expected to process or exclude from processing. */
            airlines?: components["schemas"]["RequestedAirlines"];
            /** @description Contains qualifiers that allow you to modify the price of offers. */
            fare?: components["schemas"]["RequestedFare"];
            /** @description Contains qualifiers which allow you to obtain retailing features. */
            retailing?: components["schemas"]["RequestedRetailingAttributes"];
            /** @description Contains distribution model details to determine the exchange processing path. */
            source?: components["schemas"]["RequestedSource"];
            /** @description Lists tickets whose details should be used to determine possible exchange offers. */
            tickets?: components["schemas"]["TicketToExchange"][];
            /** @description Contains configuration attributes. */
            processingOptions?: components["schemas"]["ProcessingOptions"];
        };
// ...
        FlightReshopResponse: {
            /**
             * Format: date-time
             * @description The exact point in time when the response was generated. Expressed in UTC and presented in the `YYYY-MM-DDTHH:MM:SS.SSSZ` format.
             * @example 2020-10-28T11:11:21.123Z
             */
            timestamp?: string;
            /** @description Lists all flights utilized to construct journeys of exchange offerings. */
            flights?: components["schemas"]["Flight"][];
            /** @description Lists all journeys utilized to construct exchange offerings. */
            journeys?: components["schemas"]["Journey"][];
            /**
             * Format: int32
             * @description The total number of returned offers.
             * @example 1
             */
            numberOfOffers?: number;
            /** @description Lists a selection of offers that match the requested exchange conditions. */
            offers?: components["schemas"]["Offer"][];
            /** @description Contains offer attributes for both branded and non-branded fares. */
            offerAttributes?: components["schemas"]["OfferAttributes"];
            /** @description Lists Electronic Miscellaneous Documents associated with a particular electronic flight ticket (EMD-A). */
            associatedElectronicMiscellaneousDocuments?: components["schemas"]["AssociatedElectronicMiscellaneousDocuments"][];
            /** @description Lists detailed error information. This array is not displayed in successful responses. */
            errors?: components["schemas"]["Error"][];
            /** @description Lists detailed warning information. */
            warnings?: components["schemas"]["Warning"][];
            /** @description Contains a copy of the request. */
            request?: components["schemas"]["FlightReshopRequest"];
        };
// ...
        RequestedJourney: {
            /** @description Contains a departure city or airport IATA code. */
            departureLocation: components["schemas"]["Location"];
// ...
        };
// ...
        Offer: {
// ...
            totalPriceDifference?: components["schemas"]["TotalPrice"];
// ...
            items: components["schemas"]["OfferItem"][];
            /** @description Lists references to requested additional offer solutions, based on common journey details, formerly known as brands. */
            additionalOffersRefs?: string[];
        };
// ...
        OfferItem: {
            /**
             * @description The ID of the offer item.
             * @example 123eww324-1-1
             */
            id: string;
            /** @description Lists fare details for the given offer. */
            fares: components["schemas"]["ExchangeFare"][];
        };
// ...
        ExchangeCharge: {
            /** @description Identifies the type of `grandTotal` charge. Can be `Add collect` (additional collection required to fulfill the offer), `Even` (no additional cost required to fulfill the offer), `Refund` (a refund amount will be returned when fulfilling the offer), or `Unknown` (the charge type can't be determined). */
            type?: components["schemas"]["ExchangeChargeTypeEnum"];
            /**
             * Format: number
             * @description The total fare difference.
             * @example 100.00
             */
            baseFare?: string;
            /**
             * Format: number
             * @description The total tax difference.
             * @example 8.00
             */
            totalTax?: string;
            /**
             * Format: number
             * @description The subtotal fare and tax difference, with any change fees excluded.
             * @example 100.00
             */
            subtotalBeforeFee?: string;
            /**
             * Format: number
             * @description The complete cost or refund value.
             * @example 128.00
             */
            grandTotal?: string;
            /**
             * @description The three-letter ISO 4217 currency code.
             * @example USD
             */
            currencyCode?: string;
            /**
             * Format: number
             * @description The total amount of tax on change fees.
             * @example 20.00
             */
            totalTaxOnFee?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}